+++ data/Makefile
@@ -1,0 +1,4 @@
+install:
+	apt install tesseract-ocr
+	apt install libtesseract-dev
+	apt install tesseract-ocr-slv
