+++ data/requirements.txt
@@ -1,1 +1,2 @@
 Pillow
+pytesseract
